+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- core : set max number of directional lights to 4
+
 ## [0.1.1] - 2025-05-26
 
 ### Added