+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-05-31
+
 ### Fixed
 
 - core: `Renderer` : fix not cleaning up depth texture if non-null (*critical*)
@@ -112,7 +114,8 @@ This is the first release to use a changelog.
 - Visual bug (uninitialized data) when disabling SSAO
 
 
-[Unreleased]: https://github.com/Simple-Robotics/candlewick/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/Simple-Robotics/candlewick/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/Simple-Robotics/candlewick/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/Simple-Robotics/candlewick/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/Simple-Robotics/candlewick/compare/v0.0.7...v0.1.0
 [0.0.7]: https://github.com/Simple-Robotics/candlewick/compare/v0.0.6...v0.0.7